+++ adapters/Nend/NendAdapter/GADMAdapterNend.m
@@ -6,15 +6,12 @@
 //
 
 #import "GADMAdapterNend.h"
-#import "GADMAdapterNendSetting.h"
+#import "GADMAdapterNendConstants.h"
 
 @import NendAd;
 
 @implementation GADMAdapterNendExtras
 @end
-
-static NSString *const kDictionaryKeyApiKey = @"apiKey";
-static NSString *const kDictionaryKeySpotId = @"spotId";
 
 typedef NS_ENUM(NSInteger, InterstitialVideoStatus) {
   InterstitialVideoStopped,
@@ -53,7 +50,7 @@ static GADAdSize GADSupportedAdSizeFromRequestedSize(GADAdSize gadAdSize) {
 @implementation GADMAdapterNend
 
 + (NSString *)adapterVersion {
-  return GADM_ADAPTER_NEND_VERSION;
+  return kGADMAdapterNendVersion;
 }
 
 + (Class<GADAdNetworkExtras>)networkExtrasClass {
@@ -76,11 +73,15 @@ static GADAdSize GADSupportedAdSizeFromRequestedSize(GADAdSize gadAdSize) {
 
 - (void)getInterstitial {
   id<GADMAdNetworkConnector> strongConnector = self.connector;
-  NSString *apiKey = [self getNendAdParam:kDictionaryKeyApiKey];
-  NSString *spotId = [self getNendAdParam:kDictionaryKeySpotId];
+  NSString *apiKey = [self getNendAdParam:kGADMAdapterNendApiKey];
+  NSString *spotId = [self getNendAdParam:kGADMAdapterNendSpotID];
 
   if (![self validateApiKey:apiKey spotId:spotId]) {
-    [strongConnector adapter:self didFailAd:nil];
+    NSError *error = [NSError
+        errorWithDomain:kGADMAdapterNendErrorDomain
+                   code:kGADErrorInternalError
+               userInfo:@{NSLocalizedDescriptionKey : @"SpotID and apiKey must not be nil"}];
+    [strongConnector adapter:self didFailAd:error];
     return;
   }
 
@@ -93,7 +94,7 @@ static GADAdSize GADSupportedAdSizeFromRequestedSize(GADAdSize gadAdSize) {
     self.interstitialVideo = [[NADInterstitialVideo alloc] initWithSpotId:spotId apiKey:apiKey];
     self.interstitialVideo.delegate = self;
     self.interstitialVideo.userId = extras.userId;
-    self.interstitialVideo.mediationName = @"AdMob";
+    self.interstitialVideo.mediationName = kGADMAdapterNendMediationName;
     [self.interstitialVideo loadAd];
   } else {
     self.interstitial = [NADInterstitial sharedInstance];
@@ -108,17 +109,27 @@ static GADAdSize GADSupportedAdSizeFromRequestedSize(GADAdSize gadAdSize) {
   adSize = GADSupportedAdSizeFromRequestedSize(adSize);
 
   if (GADAdSizeEqualToSize(adSize, kGADAdSizeInvalid)) {
-    [strongConnector adapter:self didFailAd:nil];
+      NSString *errorMsg =
+      [NSString stringWithFormat:@"Unable to retrieve supported ad size from GADAdSize: %@",
+       NSStringFromGADAdSize(adSize)];
+      NSError *error = [NSError errorWithDomain:kGADMAdapterNendErrorDomain
+                                           code:kGADErrorInternalError
+                                       userInfo:@{NSLocalizedDescriptionKey : errorMsg}];
+      [strongConnector adapter:self didFailAd:error];
     return;
   }
 
   self.nadView = [[NADView alloc] initWithFrame:CGRectZero];
 
-  NSString *apiKey = [self getNendAdParam:kDictionaryKeyApiKey];
-  NSString *spotId = [self getNendAdParam:kDictionaryKeySpotId];
+  NSString *apiKey = [self getNendAdParam:kGADMAdapterNendApiKey];
+  NSString *spotId = [self getNendAdParam:kGADMAdapterNendSpotID];
 
   if (![self validateApiKey:apiKey spotId:spotId]) {
-    [strongConnector adapter:self didFailAd:nil];
+    NSError *error = [NSError
+        errorWithDomain:kGADMAdapterNendErrorDomain
+                   code:kGADErrorInternalError
+               userInfo:@{NSLocalizedDescriptionKey : @"SpotID and apiKey must not be nil"}];
+    [strongConnector adapter:self didFailAd:error];
     return;
   }
 
@@ -204,7 +215,11 @@ static GADAdSize GADSupportedAdSizeFromRequestedSize(GADAdSize gadAdSize) {
 - (void)nadViewDidFailToReceiveAd:(NADView *)adView {
   NSLog(@"[nend adapter] Banner did fail to load...");
   [self.nadView pause];
-  [self.connector adapter:self didFailAd:nil];
+  NSError *error =
+      [NSError errorWithDomain:kGADMAdapterNendErrorDomain
+                          code:kGADErrorInternalError
+                      userInfo:@{NSLocalizedDescriptionKey : @"Failed to load banner ad."}];
+  [self.connector adapter:self didFailAd:error];
 }
 
 - (void)nadViewDidClickAd:(NADView *)adView {
@@ -224,7 +239,11 @@ static GADAdSize GADSupportedAdSizeFromRequestedSize(GADAdSize gadAdSize) {
   if (status == SUCCESS) {
     [strongConnector adapterDidReceiveInterstitial:self];
   } else {
-    [strongConnector adapter:self didFailAd:nil];
+    NSError *error =
+        [NSError errorWithDomain:kGADMAdapterNendErrorDomain
+                            code:kGADErrorInternalError
+                        userInfo:@{NSLocalizedDescriptionKey : @"Failed to load interstitial ad."}];
+    [strongConnector adapter:self didFailAd:error];
   }
 }