+++ adapters/Nend/NendAdapter/GADMediationAdapterNend.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2019 Google Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#import <Foundation/Foundation.h>
 @import GoogleMobileAds;
 
-/// Adapter for communicating with Facebook's Audience Network to fetch banner, interstitial,
-/// reward-based video, and native ads.
-@interface GADMAdapterFacebook : NSObject <GADMAdNetworkAdapter>
+@interface GADMediationAdapterNend : NSObject <GADMediationAdapter>
 
 @end