+++ adapters/Facebook/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Facebook Audience Network Mediation Adapter for Google Mobile Ads SDK for iOS
 
+## Version 5.2.0.2
+- Added open bidding capability to the adapter for all ad formats.
+
 ## Version 5.2.0.1
 - Updated the adapter to use the new rewarded API.
 - Now requires Google Mobile Ads SDK version 7.41.0 or higher.