+++ adapters/Nend/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Nend Adapter for Google Mobile Ads SDK for iOS
 
 ## Version 5.1.1.0
+- Updated the adapter to use the new rewarded API.
+- Now requires Google Mobile Ads SDK version 7.42.2 or higher.
 - Verified compatibility with nend SDK 5.1.1.
 - Changed condition that checking banner size.  
   Appropriate size for SmartBanner